requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
azure-core==1.9.0
azure-eventhub==5.2.0
azure-eventhub-checkpointstoreblob-aio==1.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
idna==2.10
idna-ssl==1.1.0
isodate==0.6.0
msrest==0.6.19
multidict==4.7.6
oauthlib==3.1.0
pycparser==2.20
pycurl==7.43.0
pygobject==3.22.0
python-apt==1.4.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
typing-extensions==3.7.4.3
uamqp==1.2.12
unattended-upgrades==0.1
urllib3==1.25.11
yarl==1.5.1